Refactored ensemble select dialog and made it modal
