--- frontend/src/framework/components/EnsembleSelector/index.ts
@@ -1,1 +1,0 @@
-export { EnsembleSelector } from "./ensembleSelector";
