+++ frontend/src/framework/components/SelectEnsemblesDialog/index.ts
@@ -1,0 +1,1 @@
+export { SelectEnsemblesDialog as EnsembleSelector } from "./selectEnsemblesDialog";
